pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "isplanktonbio"
version = "0.1.0"
dependencies = [
    "torch>=2.5.1",
    "torchvision",
    "torcheval",
    "faiss-cpu",
    "opencv-python",
    "matplotlib",
    "PyYAML",
    "tqdm"
]
